requirements.txt
ansible-runner==2.4.1
lockfile==0.12.2
packaging==25.0
pexpect==4.9.0
ptyprocess==0.7.0
python-daemon==3.1.2
PyYAML==6.0.2
